+++ src/happy/readers/_matlab_reader.py
@@ -7,7 +7,7 @@ import scipy.io as sio
 
 from happy.data import HappyData
 from ._happydata_reader import HappyDataReader
-from seppl.placeholders import expand_placeholders
+from seppl.variables import expand_variables
 
 
 class MatlabReader(HappyDataReader):
@@ -26,7 +26,7 @@ class MatlabReader(HappyDataReader):
 
     def _get_sample_ids(self) -> List[str]:
         sample_ids = []
-        base_dir = expand_placeholders(self.base_dir)
+        base_dir = expand_variables(self.base_dir)
         for f in os.listdir(base_dir):
             if f.endswith(".mat"):
                 sample_path = os.path.join(base_dir, f)
@@ -57,7 +57,7 @@ class MatlabReader(HappyDataReader):
         metadata_dict = {}
 
         # read matlab
-        filename = os.path.join(expand_placeholders(self.base_dir), "%s.%s.mat" % (sample_id, region_name))
+        filename = os.path.join(expand_variables(self.base_dir), "%s.%s.mat" % (sample_id, region_name))
         mat_file = sio.loadmat(filename)
 
         # data