+++ src/happy/writers/_image_writer.py
@@ -6,7 +6,7 @@ from typing import Tuple, Optional
 from happy.data import HappyData, DataManager
 from happy.data.normalization import SimpleNormalization
 from ._happydata_writer import HappyDataWriterWithOutputPattern, HappyDataWriterWithNormalization, PH_BASEDIR, PH_SAMPLEID, PH_REPEAT
-from seppl.placeholders import expand_placeholders
+from seppl.variables import expand_variables
 
 
 class ImageWriter(HappyDataWriterWithOutputPattern, HappyDataWriterWithNormalization):
@@ -85,7 +85,7 @@ class ImageWriter(HappyDataWriterWithOutputPattern, HappyDataWriterWithNormaliza
             self.logger().info(msg)
         sample_id = happy_data.sample_id
         region_id = happy_data.region_id
-        base_dir = expand_placeholders(self.base_dir)
+        base_dir = expand_variables(self.base_dir)
         self.logger().info("Creating dir: %s" % base_dir)
         os.makedirs(base_dir, exist_ok=True)
         # output image