+++ CHANGES.rst
@@ -9,6 +9,7 @@ Changelog
 - requiring seppl>=0.2.13 now
 - added basic placeholder support (readers/writers)
 - using underscores now instead of dashes in dependencies (`setup.py`)
+- requiring seppl>=0.3.1 now and switched to using seppl.variables
 
 
 0.0.3 (2025-03-07)